# ServiceApp.ApiClient/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ServiceApp.API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Customer/GetCustomer": {
      "get": {
        "tags": [
          "Customer"
        ],
        "operationId": "GetCustomer",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/UserDto"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/api/Customer/GetAllCustomers": {
      "get": {
        "tags": [
          "Customer"
        ],
        "operationId": "GetAllCustomers",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/CustomerDto"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/api/Customer/CreateCustomer": {
      "post": {
        "tags": [
          "Customer"
        ],
        "operationId": "CreateCustomer",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CustomerDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Customer"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/api/Role/GetRoles": {
      "get": {
        "tags": [
          "Role"
        ],
        "operationId": "GetRoles",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Role"
              }
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
    "/api/User/GetUserByEmail": {
      "get": {
        "tags": [
          "User"
        ],
        "operationId": "GetUserByEmail",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "email",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/UserDto"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Comment": {
      "type": "object",
      "properties": {
        "id": {
          "format": "uuid",
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "createdDate": {
          "format": "date-time",
          "type": "string"
        },
        "author": {
          "$ref": "#/definitions/User"
        },
        "ticket": {
          "$ref": "#/definitions/Ticket"
        }
      }
    },
    "CommentDto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "uuid",
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "createdDate": {
          "format": "date-time",
          "type": "string"
        },
        "author": {
          "$ref": "#/definitions/UserDto"
        }
      }
    },
...
    "Customer": {
      "required": [
        "email",
        "firstName",
        "id",
        "lastName"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "uuid",
          "type": "string"
        },
        "firstName": {
          "maxLength": 100,
          "minLength": 2,
          "type": "string"
        },
        "lastName": {
          "maxLength": 100,
          "minLength": 2,
          "type": "string"
        },
        "email": {
          "format": "email",
          "minLength": 1,
          "type": "string"
        },
        "phoneNumber": {
          "format": "tel",
          "type": "string"
        },
        "passwordHash": {
          "format": "byte",
          "type": "string"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "passwordSalt": {
          "format": "byte",
          "type": "string"
        },
        "isActive": {
          "type": "boolean"
        },
        "roles": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Role"
          }
        },
        "resetPasswordToken": {
          "type": "string"
        },
        "resetTokenExpires": {
          "format": "date-time",
          "type": "string"
        },
        "refreshToken": {
          "type": "string"
        },
        "refreshTokenCreated": {
          "format": "date-time",
          "type": "string"
        },
        "refreshTokenExpires": {
          "format": "date-time",
          "type": "string"
        },
        "ticketsToProcess": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Ticket"
          }
        },
        "ticketsCreated": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Ticket"
          }
        },
        "jiraCredentials": {
          "$ref": "#/definitions/JiraCredentials"
        },
        "customerId": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
    "CustomerDto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "uuid",
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "isActive": {
          "type": "boolean"
        },
        "roles": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/RoleDto"
          }
        }
      }
    },
    "JiraCredentials": {
      "type": "object",
      "properties": {
        "id": {
          "format": "uuid",
          "type": "string"
        },
        "user": {
          "$ref": "#/definitions/User"
        },
        "password": {
          "type": "string"
        },
        "username": {
          "type": "string"
        }
      }
    },
    "JiraCredentialsDto": {
      "type": "object",
      "properties": {
        "password": {
          "type": "string"
        },
        "username": {
          "type": "string"
        }
      }
    },
...
    "Role": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "roleName": {
          "type": "string"
        },
        "users": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/User"
          }
        }
      }
    },
    "RoleDto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "roleName": {
          "type": "string"
        }
      }
    },
    "Ticket": {
      "required": [
        "description",
        "title"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "uuid",
          "type": "string"
        },
        "title": {
          "maxLength": 100,
          "minLength": 2,
          "type": "string"
        },
        "description": {
          "maxLength": 1000,
          "minLength": 2,
          "type": "string"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "modifiedAt": {
          "format": "date-time",
          "type": "string"
        },
        "dueTime": {
          "format": "date-time",
          "type": "string"
        },
        "responsibleUser": {
          "$ref": "#/definitions/User"
        },
        "createdBy": {
          "$ref": "#/definitions/User"
        },
        "modifiedBy": {
          "$ref": "#/definitions/User"
        },
        "jiraTicketId": {
          "type": "string"
        },
        "ticketStatus": {
          "$ref": "#/definitions/TicketStatus"
        },
        "comments": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Comment"
          }
        },
        "priority": {
          "$ref": "#/definitions/TicketPriority"
        }
      }
    },
    "TicketDto": {
      "required": [
        "title"
      ],
      "type": "object",
      "properties": {
        "title": {
          "minLength": 1,
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "jiraTicketId": {
          "type": "string"
        },
        "id": {
          "format": "uuid",
          "type": "string"
        },
        "comments": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CommentDto"
          }
        },
        "ticketStatus": {
          "$ref": "#/definitions/TicketStatus"
        },
        "modifiedAt": {
          "format": "date-time",
          "type": "string"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "dueTime": {
          "format": "date-time",
          "type": "string"
        },
        "priority": {
          "$ref": "#/definitions/TicketPriority"
        },
        "modifiedById": {
          "format": "uuid",
          "type": "string"
        },
        "responsibleUser": {
          "$ref": "#/definitions/UserDto"
        },
        "createdBy": {
          "$ref": "#/definitions/UserDto"
        }
      }
    },
    "TicketPriority": {
      "enum": [
        "Lowest",
        "Low",
        "Normal",
        "High",
        "Highest"
      ],
      "type": "string"
    },
    "TicketStatus": {
      "enum": [
        "Pending",
        "WorkInProgress",
        "Open",
        "ClosedComplete"
      ],
      "type": "string"
    },
...
    "User": {
      "required": [
        "email",
        "firstName",
        "id",
        "lastName"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "uuid",
          "type": "string"
        },
        "firstName": {
          "maxLength": 100,
          "minLength": 2,
          "type": "string"
        },
        "lastName": {
          "maxLength": 100,
          "minLength": 2,
          "type": "string"
        },
        "email": {
          "format": "email",
          "minLength": 1,
          "type": "string"
        },
        "phoneNumber": {
          "format": "tel",
          "type": "string"
        },
        "passwordHash": {
          "format": "byte",
          "type": "string"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "passwordSalt": {
          "format": "byte",
          "type": "string"
        },
        "isActive": {
          "type": "boolean"
        },
        "roles": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Role"
          }
        },
        "resetPasswordToken": {
          "type": "string"
        },
        "resetTokenExpires": {
          "format": "date-time",
          "type": "string"
        },
        "refreshToken": {
          "type": "string"
        },
        "refreshTokenCreated": {
          "format": "date-time",
          "type": "string"
        },
        "refreshTokenExpires": {
          "format": "date-time",
          "type": "string"
        },
        "ticketsToProcess": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Ticket"
          }
        },
        "ticketsCreated": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Ticket"
          }
        },
        "jiraCredentials": {
          "$ref": "#/definitions/JiraCredentials"
        }
      }
    },
    "UserDto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "uuid",
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        },
        "isActive": {
          "type": "boolean"
        },
        "roles": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/RoleDto"
          }
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "ticketsCreated": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/TicketDto"
          }
        },
        "jiraCredentials": {
          "$ref": "#/definitions/JiraCredentialsDto"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "oauth2": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Standard Authorization header using the Bearer scheme (\"bearer {token}\")"
    }
  }
}