# ServiceApp.Models/Entities/TicketStatus.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace ServiceApp.Models.Entities
{
    public enum TicketStatus
    {
        Pending=0,
        WorkInProgress=1,
        Open=2,
        ClosedComplete=3
    }
}